refactor: split learner chat hook helpers
Changed:
Moved pure helper logic and local hook types out of the learner chat hook into dedicated modules.

Benefit:
The learner stream hook is easier to review and change without altering runtime behavior.
